For issue #58.
Moves C-style memory management code out from Settings and into its own files.
